Add a fuzzer against a bitvec, to compare correctness
