+++ fuzz/Cargo.toml
@@ -26,3 +26,9 @@ name = "fuzz_ops"
 path = "fuzz_targets/fuzz_ops.rs"
 test = false
 doc = false
+
+[[bin]]
+name = "against_bitvec"
+path = "fuzz_targets/against_bitvec.rs"
+test = false
+doc = false
